+++ caikit_nlp/modules/text_generation/peft_tgis_remote.py
@@ -18,6 +18,9 @@ prompt vectors in TGIS generation requests.
 from typing import Iterable, List, Optional, Tuple, Union
 import os
 
+# Third Party
+import numpy as np
+
 # First Party
 from caikit.core import ModuleBase, ModuleConfig, ModuleSaver, modules
 from caikit.core.module_backends import BackendBase, backend_types
@@ -175,7 +178,7 @@ class PeftPromptTuningTGIS(ModuleBase):
         top_p: Optional[float] = 1.0,
         typical_p: Optional[float] = 1.0,
         temperature: Optional[float] = 1.0,
-        seed: Optional[int] = None,
+        seed: Optional[np.uint64] = None,
         repetition_penalty: Optional[float] = 1.0,
         max_time: Optional[float] = None,
         exponential_decay_length_penalty: Optional[
@@ -231,7 +234,7 @@ class PeftPromptTuningTGIS(ModuleBase):
         top_p: Optional[float] = 1.0,
         typical_p: Optional[float] = 1.0,
         temperature: Optional[float] = 1.0,
-        seed: Optional[int] = None,
+        seed: Optional[np.uint64] = None,
         repetition_penalty: Optional[float] = 1.0,
         max_time: Optional[float] = None,
         exponential_decay_length_penalty: Optional[
